1282 Update for new Docket PDF Options
