+++ shared/src/business/useCases/generatePdfFromHtmlInteractor.js
@@ -1,7 +1,3 @@
-const {
-  generatePdfFromHtml,
-} = require('../useCaseHelper/pdf/generatePdfFromHtml');
-
 /**
  * generatePdfFromHtmlInteractor
  *
@@ -19,11 +15,66 @@ exports.generatePdfFromHtmlInteractor = async ({
   docketNumber,
   headerHtml,
 }) => {
-  return generatePdfFromHtml({
-    applicationContext,
-    contentHtml,
-    displayHeaderFooter,
-    docketNumber,
-    headerHtml,
-  });
+  let browser = null;
+  let result = null;
+
+  try {
+    applicationContext.logger.time('Generating PDF From HTML');
+    const chromium = applicationContext.getChromium();
+
+    browser = await chromium.puppeteer.launch({
+      args: chromium.args,
+      defaultViewport: chromium.defaultViewport,
+      executablePath: await chromium.executablePath,
+      headless: true,
+    });
+
+    let page = await browser.newPage();
+
+    await page.setContent(contentHtml);
+
+    const headerTemplate = `
+      <!doctype html>
+      <html>
+        <head>
+        </head>
+        <body style="margin: 0px;">
+          <div style="font-size: 8px; font-family: sans-serif; width: 100%; margin: 0px 40px; margin-top: 25px;">
+            <div style="font-size: 8px; font-family: sans-serif; float: right;">
+              Page <span class="pageNumber"></span>
+              of <span class="totalPages"></span>
+            </div>
+            <div style="float: left">
+              ${headerHtml ? headerHtml : `Docket Number: ${docketNumber}`}
+            </div>
+          </div>
+        </body>
+      </html>
+    `;
+
+    const footerTemplate = `
+      <div style="font-size:8px !important; color:#000; text-align:center; width:100%; margin-bottom:5px;">Printed <span class="date"></span></div>
+    `;
+
+    result = await page.pdf({
+      displayHeaderFooter,
+      footerTemplate: footerTemplate,
+      format: 'Letter',
+      headerTemplate: headerTemplate,
+      margin: {
+        bottom: '100px',
+        top: '80px',
+      },
+      printBackground: true,
+    });
+  } catch (error) {
+    applicationContext.logger.error(error);
+    throw error;
+  } finally {
+    if (browser !== null) {
+      await browser.close();
+    }
+  }
+  applicationContext.logger.timeEnd('Generating PDF From HTML');
+  return result;
 };
