+++ web-client/src/presenter/sequences/downloadBatchOfTrialSessionSequence.js
@@ -1,14 +1,11 @@
 import { downloadBatchOfTrialSessionAction } from '../actions/TrialSession/downloadBatchOfTrialSessionAction';
 import { getCalendaredCasesForTrialSessionAction } from '../actions/TrialSession/getCalendaredCasesForTrialSessionAction';
 import { getCaseAction } from '../actions/getCaseAction';
-import { getCaseDeadlinesForCaseAction } from '../actions/CaseDeadline/getCaseDeadlinesForCaseAction';
 import { getTrialSessionDetailsAction } from '../actions/TrialSession/getTrialSessionDetailsAction';
 import { hasAnotherCalendaredCaseAction } from '../actions/TrialSession/hasAnotherCalendaredCaseAction';
 import { isTrialSessionCalendaredAction } from '../actions/TrialSession/isTrialSessionCalendaredAction';
-import { printDocketRecordAction } from '../actions/printDocketRecordAction';
 import { setCaseAction } from '../actions/setCaseAction';
-import { setDefaultDocketRecordSortAction } from '../actions/DocketRecord/setDefaultDocketRecordSortAction';
-import { stashDocketRecordHtmlAction } from '../actions/TrialSession/stashDocketRecordHtmlAction';
+import { stashCaseDetailByCaseIdAction } from '../actions/TrialSession/stashCaseDetailByCaseIdAction';
 
 let sequenceHead, sequenceNode, sequenceYes;
 
@@ -16,14 +13,7 @@ sequenceHead = sequenceNode = [];
 
 // TODO Fix to have docket record generation server side then remove this.
 for (let i = 0; i < 1000; i++) {
-  sequenceYes = [
-    getCaseAction,
-    setCaseAction,
-    getCaseDeadlinesForCaseAction,
-    setDefaultDocketRecordSortAction,
-    printDocketRecordAction,
-    stashDocketRecordHtmlAction,
-  ];
+  sequenceYes = [getCaseAction, setCaseAction, stashCaseDetailByCaseIdAction];
   sequenceNode.push(hasAnotherCalendaredCaseAction);
   sequenceNode.push({
     no: [],
